LayoutTests/fast/css/getComputedStyle/computed-style-font-family-expected.txt

Font attributes. The font-family should list three families:

font-family: monospace, 'Lucida Grande', sans-serif;
font-size: 13px;
font-style: normal;
font-variant: normal;
font-weight: normal;
